fix: 修复字体未按照10%的比例进行缩放
Description: 旧版代码按字体大小加减1进行缩放，未按10%的比例进行缩放。
修改字体大小为浮点类型，调整缩放计算方式，允许更高精度缩放文本大小。

Log: 修复字体未按照10%的比例进行缩放
Influence: 底栏的缩放比例展示及字体缩放

+++ src/common/settings.h
@@ -69,7 +69,7 @@ public:
 
 signals:
     void sigAdjustFont(QString name);
-    void sigAdjustFontSize(int fontSize);
+    void sigAdjustFontSize(qreal fontSize);
     void sigAdjustTabSpaceNumber(int number);
     void sigAdjustWordWrap(bool enable);
     void sigAdjustBookmark(bool enable);
